Tenant API & Panel Organization
This is a larger body of work than I'd normally like to commit, but it's all tied together in a way that makes it hard to split.

The goal here was to enable Tenant-Context API routes, using custom guards based on the panel name (segment 2 in the path).

This tied into a renaming of the two initial panels from Admin / Manage to Account / Operate (https://somedomain.com/account for subscription management, https://somedomain.com/operate for app functions).